requirements.txt
llmsherpa==0.1.4
neo4j==5.27.0
python-dotenv==1.0.1

pymupdf==1.25.1
pymupdf4llm==0.0.17
pdfplumber==0.11.5
pdfminer.six==20231228

camelot-py==1.0.0
langchain==0.3.14

markdown-it-py==3.0.0


pydantic==2.10.4


langchain_neo4j==0.2.0
langchain_openai==0.3.1
langchain_community==0.3.14

openai==1.58.1
